move components to newfile
